DinoAIModel/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
Pillow==11.1.0
python-multipart==0.0.20
timm